attention: delete FlashAttention wrapper; move decomposed oracles to tests
FlashAttention is gone with no replacement helper: Core's fused branch
covers its behavior through InternalFusedOpCaller, and consumers needing
the bare pre-projected fused-or-decomposed op call
graph.InternalFusedOpCaller + BackendFusedScaledDotProductAttention
directly. naiveCausalAttention has no runtime caller after the deletion,
so it joins naiveGQAReference and repeatKVHeads as test-only decomposed
oracles in fusion_test.go. flash_bench_test.go's attentionStep is stubbed
to the decomposed path (FlashAttention ref removed; TODO Task 7 to rework
with WithFusion).
